requirements.txt
streamlit>=1.28.0
langchain_groq>=0.0.3
langchain_community>=0.0.10
pypdf>=3.17.0
sentence_transformers>=2.2.2
python-dotenv>=1.0.0
chromadb>=0.4.18
langchain-chroma>=0.0.1
langchain-text-splitters>=0.0.1
rank_bm25>=0.2.2
langchain-huggingface>=0.0.1
pandas>=2.0.0
openpyxl>=3.1.2
python-docx>=1.1.0
Pillow>=10.0.0